+++ src/String.cpp
@@ -173,12 +173,12 @@ template<> unsigned long Vaca::convert_to(const String& from)
 
 template<> float Vaca::convert_to(const String& from)
 {
-  return std::wcstod(from.c_str(), NULL);
+  return (float)std::wcstod(from.c_str(), NULL);
 }
 
 template<> double Vaca::convert_to(const String& from)
 {
-  return (float)std::wcstod(from.c_str(), NULL);
+  return std::wcstod(from.c_str(), NULL);
 }
 
 template<> String Vaca::convert_to(const char* const& from)
